+++ src/pages/libraryPage/EditBookCollectionPageStyle.ts
@@ -22,12 +22,12 @@ export const styles = StyleSheet.create({
     ...Font.Label.XMedium,
     color: "#394149",
   },
-  collectionList: {
-    height: 350,
-  },
-  collectionWrap: {
+  collectionMinusWrap: {
     marginBottom: 16,
     marginHorizontal: 13,
+  },
+  collectionList: {
+    height: 350,
   },
   collectionPlusWrap: {
     alignItems: "center",
@@ -35,61 +35,5 @@ export const styles = StyleSheet.create({
   flatListContent: {
     marginBottom: 16,
     justifyContent: "center",
-  },
-  collectionItem: {
-    width: 136,
-    marginHorizontal: 3,
-  },
-  collectionPlusItem: {
-    width: 164,
-    marginHorizontal: 12,
-    marginBottom: 6,
-  },
-  collectionCover: {
-    position: "absolute",
-    borderRadius: 3,
-    backgroundColor: "rgba(0, 0, 0, 0.6)",
-    width: "100%",
-    height: "100%",
-    zIndex: 1,
-  },
-  collectionImages: {
-    position: "relative",
-  },
-  minusIcon: {
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: [{ translateX: -34.565 }, { translateY: -34.565 }],
-    zIndex: 2,
-  },
-  imageGrid: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    justifyContent: "center",
-    alignItems: "center",
-    marginVertical: 4,
-    zIndex: 0,
-  },
-  collectionImage: {
-    width: 60,
-    height: 91.33,
-    margin: 2,
-  },
-  collectionPlusImage: {
-    width: 74,
-    height: 112,
-    margin: 2,
-  },
-  textWrap: {
-    marginVertical: 4,
-  },
-  collectionTitleText: {
-    ...Font.Label.SemiMedium,
-    color: "#3F474F",
-  },
-  collectionNumText: {
-    ...Font.Paragraph.SemiMedium,
-    color: "#3F474F",
   },
 });
